refactor: useCardRegistrationFlow 훅을 index.ts로 이동하고 useFocusControl 훅 추가
